chore: no 4 cols design
