+++ src/routes/(main)/agent/features/Conversation/RightPanel/index.tsx
@@ -7,6 +7,7 @@ import NavHeader from '@/features/NavHeader';
 import RightPanel from '@/features/RightPanel';
 import { useGlobalStore } from '@/store/global';
 
+import AgentDocumentEditorPanel from './AgentDocumentEditorPanel';
 import AgentSummary from './AgentSummary';
 import ProgressSection from './ProgressSection';
 import ResourcesSection from './ResourcesSection';
@@ -19,34 +20,42 @@ interface AgentWorkspaceRightPanelProps {
 const AgentWorkspaceRightPanel = memo<AgentWorkspaceRightPanelProps>(
   ({ onSelectDocument, selectedDocumentId }) => {
     const toggleRightPanel = useGlobalStore((s) => s.toggleRightPanel);
+    const isDocumentMode = Boolean(selectedDocumentId);
 
     return (
-      <RightPanel defaultWidth={360} maxWidth={520} minWidth={300}>
-        <Flexbox height={'100%'} width={'100%'}>
-          <NavHeader
-            showTogglePanelButton={false}
-            right={
-              <ActionIcon
-                icon={PanelRightCloseIcon}
-                size={DESKTOP_HEADER_ICON_SIZE}
-                onClick={() => toggleRightPanel(false)}
-              />
-            }
-            style={{
-              paddingBlock: 8,
-              paddingInline: 8,
-              position: 'absolute',
-            }}
+      <RightPanel defaultWidth={360} maxWidth={720} minWidth={300}>
+        {isDocumentMode ? (
+          <AgentDocumentEditorPanel
+            selectedDocumentId={selectedDocumentId}
+            onClose={() => onSelectDocument(null)}
           />
-          <Flexbox gap={8} height={'100%'} style={{ overflowY: 'auto' }} width={'100%'}>
-            <AgentSummary />
-            <ProgressSection />
-            <ResourcesSection
-              selectedDocumentId={selectedDocumentId}
-              onSelectDocument={onSelectDocument}
+        ) : (
+          <Flexbox height={'100%'} width={'100%'}>
+            <NavHeader
+              showTogglePanelButton={false}
+              right={
+                <ActionIcon
+                  icon={PanelRightCloseIcon}
+                  size={DESKTOP_HEADER_ICON_SIZE}
+                  onClick={() => toggleRightPanel(false)}
+                />
+              }
+              style={{
+                paddingBlock: 8,
+                paddingInline: 8,
+                position: 'absolute',
+              }}
             />
+            <Flexbox gap={8} height={'100%'} style={{ overflowY: 'auto' }} width={'100%'}>
+              <AgentSummary />
+              <ProgressSection />
+              <ResourcesSection
+                selectedDocumentId={selectedDocumentId}
+                onSelectDocument={onSelectDocument}
+              />
+            </Flexbox>
           </Flexbox>
-        </Flexbox>
+        )}
       </RightPanel>
     );
   },
